single activity & itinerary folder hierarchy
