+++ frontend/src/components/activity/activitySingle/ActivityDetails.js
@@ -1,24 +1,24 @@
-import ItineraryMainInformation from "./ItineraryMainInformation";
-// import OthersInformation from "./OthersInformation";
-// import Overview from "./Overview";
-import MapComponent from "../common/MapComponent";
-// import Rating from "./Rating";
-// import TourSingleSidebar from "./TourSingleSidebar";
-// import Gallery1 from "./Gallery1";
-// import DateCalender from "./DateCalender";
-import ReviewBox from "../common/ReviewBox";
-import ItineraryReviews from "./ItineraryReviews"; 
+import ActivityMainInformation from "./ActivityMainInformation";
+import OthersInformation from "./OthersInformation";
+import Overview from "./Overview";
+import MapComponent from "../../common/MapComponent";
+import Rating from "./Rating";
+import TourSingleSidebar from "./TourSingleSidebar";
+import Gallery1 from "./Gallery1";
+import DateCalender from "./DateCalender";
+import ReviewBox from "../../common/ReviewBox";
+import ActivityReviews from "./ActivityReviews"; 
 
-export default function ItineraryDetails({ itinerary }) {
-  if (!itinerary) return <div>Itinerary not found.</div>;
-  const itineraryId = itinerary._id;
+export default function ActivityDetails({ activity }) {
+  if (!activity) return <div>Activity not found.</div>;
+  const activityId = activity._id;
 
   return (
     <>
       <section className="">
         <div className="container">
-          <ItineraryMainInformation itinerary={itinerary} />
-          {/* <Gallery1 /> */}
+          <ActivityMainInformation activity={activity} />
+          <Gallery1 />
         </div>
       </section>
 
@@ -27,15 +27,12 @@ export default function ItineraryDetails({ itinerary }) {
           <div className="row y-gap-30 justify-between">
             <div className="col-lg-8">
               <div className="row y-gap-20 justify-between items-center layout-pb-md">
-                {/* <OthersInformation /> */}
+                <OthersInformation />
               </div>
 
-              {/* <Overview /> */}
+              <Overview />
 
               <div className="line mt-60 mb-60"></div>
-
-              <h2 className="text-30 mt-60 mb-30">Road Map</h2>
-              {/* <Roadmap2 />   */}
 
               <h2 className="text-30 mt-60 mb-30">Tour Map</h2>
               <div className="mapTourSingle">
@@ -45,17 +42,17 @@ export default function ItineraryDetails({ itinerary }) {
               <div className="line mt-60 mb-60"></div>
 
               <h2 className="text-30">Availability Calendar</h2>
-              {/* <DateCalender /> */}
+              <DateCalender />
 
               <div className="line mt-60 mb-60"></div>
 
               <h2 className="text-30">Customer Reviews</h2>
 
               <div className="mt-30">
-                {/* <Rating /> */}
+                <Rating />
               </div>
 
-              <ItineraryReviews itineraryId={itineraryId} />
+              <ActivityReviews activityId={activityId} />
 
               {/* <button className="button -md -outline-accent-1 text-accent-1 mt-30">
                 See more reviews
@@ -64,13 +61,15 @@ export default function ItineraryDetails({ itinerary }) {
 
               <div className="line mt-60 mb-60"></div>
 
-              <ReviewBox id={itineraryId} type="itinerary"/>
+              <ReviewBox id={activityId} type="activities"/>
+
+              <div className="line mt-60 mb-60"></div>
 
             </div>
 
             <div className="col-lg-4">
               <div className="d-flex justify-end js-pin-content">
-                {/* <TourSingleSidebar /> */}
+                <TourSingleSidebar />
               </div>
             </div>
           </div>
