+++ frontend/src/components/activity/activitySingle/Reviews.jsx
@@ -1,6 +1,6 @@
-import { reviews } from "../../data/tourSingleContent";
+import { reviews } from "../../../data/tourSingleContent";
 import React from "react";
-import Stars from "../common/Stars";
+import Stars from "../../common/Stars";
 
 export default function Reviews() {
   return (
