+++ frontend/src/components/activity/activitySingle/Roadmap2.jsx
@@ -1,4 +1,4 @@
-import { roadmapData2 } from "../../data/tourSingleContent";
+import { roadmapData2 } from "../../../data/tourSingleContent";
 import React, { useState } from "react";
 
 export default function RoadMap2() {
