+++ frontend/src/components/itinerary/itinerarySingle/ItineraryDetails.js
@@ -1,24 +1,24 @@
-import ActivityMainInformation from "./ActivityMainInformation";
-import OthersInformation from "./OthersInformation";
-import Overview from "./Overview";
-import MapComponent from "../common/MapComponent";
-import Rating from "./Rating";
-import TourSingleSidebar from "./TourSingleSidebar";
-import Gallery1 from "./Gallery1";
-import DateCalender from "./DateCalender";
-import ReviewBox from "../common/ReviewBox";
-import ActivityReviews from "./ActivityReviews"; 
+import ItineraryMainInformation from "./ItineraryMainInformation";
+// import OthersInformation from "./OthersInformation";
+// import Overview from "./Overview";
+import MapComponent from "../../common/MapComponent";
+// import Rating from "./Rating";
+// import TourSingleSidebar from "./TourSingleSidebar";
+// import Gallery1 from "./Gallery1";
+// import DateCalender from "./DateCalender";
+import ReviewBox from "../../common/ReviewBox";
+import ItineraryReviews from "./ItineraryReviews"; 
 
-export default function ActivityDetails({ activity }) {
-  if (!activity) return <div>Activity not found.</div>;
-  const activityId = activity._id;
+export default function ItineraryDetails({ itinerary }) {
+  if (!itinerary) return <div>Itinerary not found.</div>;
+  const itineraryId = itinerary._id;
 
   return (
     <>
       <section className="">
         <div className="container">
-          <ActivityMainInformation activity={activity} />
-          <Gallery1 />
+          <ItineraryMainInformation itinerary={itinerary} />
+          {/* <Gallery1 /> */}
         </div>
       </section>
 
@@ -27,12 +27,15 @@ export default function ActivityDetails({ activity }) {
           <div className="row y-gap-30 justify-between">
             <div className="col-lg-8">
               <div className="row y-gap-20 justify-between items-center layout-pb-md">
-                <OthersInformation />
+                {/* <OthersInformation /> */}
               </div>
 
-              <Overview />
+              {/* <Overview /> */}
 
               <div className="line mt-60 mb-60"></div>
+
+              <h2 className="text-30 mt-60 mb-30">Road Map</h2>
+              {/* <Roadmap2 />   */}
 
               <h2 className="text-30 mt-60 mb-30">Tour Map</h2>
               <div className="mapTourSingle">
@@ -42,17 +45,17 @@ export default function ActivityDetails({ activity }) {
               <div className="line mt-60 mb-60"></div>
 
               <h2 className="text-30">Availability Calendar</h2>
-              <DateCalender />
+              {/* <DateCalender /> */}
 
               <div className="line mt-60 mb-60"></div>
 
               <h2 className="text-30">Customer Reviews</h2>
 
               <div className="mt-30">
-                <Rating />
+                {/* <Rating /> */}
               </div>
 
-              <ActivityReviews activityId={activityId} />
+              <ItineraryReviews itineraryId={itineraryId} />
 
               {/* <button className="button -md -outline-accent-1 text-accent-1 mt-30">
                 See more reviews
@@ -61,15 +64,13 @@ export default function ActivityDetails({ activity }) {
 
               <div className="line mt-60 mb-60"></div>
 
-              <ReviewBox id={activityId} type="activities"/>
-
-              <div className="line mt-60 mb-60"></div>
+              <ReviewBox id={itineraryId} type="itinerary"/>
 
             </div>
 
             <div className="col-lg-4">
               <div className="d-flex justify-end js-pin-content">
-                <TourSingleSidebar />
+                {/* <TourSingleSidebar /> */}
               </div>
             </div>
           </div>
