+++ frontend/src/components/itinerary/itinerarySingle/ItineraryMainInformation.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Stars from "../common/Stars";
+import Stars from "../../common/Stars";
 
 export default function ItineraryMainInformation({ itinerary }) {
   return (
